requirements.txt
gitdb==4.0.10
GitPython==3.1.31
grad-cam==1.4.6
huggingface-hub==0.14.1
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
ipykernel==6.23.1
ipython==8.14.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
triton==2.0.0
ttach==0.0.3
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.3
wcwidth==0.2.6
zipp==3.15.0
dill